Add solution for stack_based_buffer_overflow_1.c
First of all, this script should be compiled with -fno-stack-protector,
assuming one is using gcc. This flag disables the buffer-overflow
protection (aka canary).

The solution is: 0000000000DCBA

DCBA (0x44 0x43 0x42 0x41) instead of ABCD (0x41 0x42 0x43 0x44)
because, in this case, we're assuming little-endiannes.

There is no way to be sure if the buffer is going to be placed after or
before the integer, but the compiler may place it after the integer in
order to avoid buffer-overflow exploitation. This protection seems to
be disabled when -fno-stack-protector is used.